Clear up some warnings

+++ starterCode/src/engine/DisplayObjectContainer.h
@@ -19,7 +19,7 @@ public:
 	void addChild(DisplayObject* child);
 	void removeImmediateChild(DisplayObject* child);
 	void removeImmediateChild(std::string id);
-	void removeChild(int index);
+	void removeChild(size_t index);
 	void removeThis();
 	
 	virtual int numChildren();
